add various options
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm/trunk@3003 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ regress/config.sh
@@ -1,51 +1,100 @@
 #!/bin/sh
 
 PROGNAME=`basename $0`
+remove_environment=true
+hostname=
+BACKEND_PORT=50602
+GFMD_PORT=50601
+GFSD_PORT=50600
+hostname=
+
+usage()
+{
+	echo >&2 "Usage: $PROGNAME [<options>] [<regression_test>]"
+	echo >&2 "mandatory options:"
+	echo >&2 "	--prefix <installatioin_prefix>"
+	echo >&2 "	--config-prefix <config_prefix>"
+	echo >&2 "	-b <backend_type>"
+	echo >&2 "options:"
+	echo >&2 "	--help				: print this message"
+	echo >&2 "	-k				: keep configuration"
+	echo >&2 "	-h <hostname>"
+	echo >&2 "	-p <metadata_backend_port>"
+	echo >&2 "	-m <gfmd_port>"
+	echo >&2 "	-s <gfsd_port>"
+	exit 2
+}
+
+while	case $1 in
+	--prefix)
+		prefix=${2?"$PROGNAME: --prefix option requires <installation_prefix> argument"}
+		shift; true;;
+	--config-prefix)
+		CONFIG_PREFIX=${2?"$PROGNAME: --config-prefix option requires <config_prefix> argument"}
+		shift; true;;
+	-b)	BACKEND_TYPE=${2?"$PROGNAME: -b option requires <backend> argument"}
+		shift; true;;
+	--help)	usage;;
+	-k)	remove_environment=true; true;;
+	-h)	hostname="${2?"$PROGNAME: -h option requires <hostname> argument"}
+		hostname="-h $hostname"
+		shift; true;;
+	-p)	BACKEND_PORT="${2?"$PROGNAME: -p option requires <metadata_backend_port> argument"}
+		shift; true;;
+	-m)	GFMD_PORT="${2?"$PROGNAME: -m option requires <gfmd_port> argument"}
+		shift; true;;
+	-s)	GFSD_PORT="${2?"$PROGNAME: -s option requires <gfsd_port> argument"}
+		shift; true;;
+	-*)	echo >&2 "$PROGNAME: unknown option $1"
+		usage;;
+	*)	false;;
+	esac
+do
+	shift
+done
+
+case $# in
+0)	set ./regress.sh;;
+esac
 
 if [ -z "$prefix" ]; then
-	echo $PROGNAME: '${prefix} is not set' >&2
+	echo >&2 $PROGNAME: '"--prefix <installation_prefix>" option is not set, aborted'
+	exit 1
+elif [ ! -d "$prefix" ]; then
+	echo >&2 $PROGNAME: "$prefix is not a directory, aborted"
 	exit 1
 fi
-config_dir="${prefix}/share/gfarm/config"
-
-. $config_dir/config-gfarm.sysdep
-
-case $1 in
--b)	BACKEND_TYPE=${2?"$PROGNAME: -b option requires <backend> argument"}
-	shift 2;;
-*)	BACKEND_TYPE=postgresql;;
-esac
-case $# in
-1)	regress=./regress.sh;;
-2)	regress=$2;;
-*)	echo "Usage: $PROGNAME <config_prefix> [<regression_test>]" >&2
-	exit 2;;
-esac
-CONFIG_PREFIX=$1
-
+if [ -z "$CONFIG_PREFIX" ]; then
+	echo >&2 $PROGNAME: '"--config-prefix <config_prefix>" option is necessary, aborted'
+	exit 1
+fi
 if [ -d "$CONFIG_PREFIX" ] || [ -f "$CONFIG_PREFIX" ]; then
-	echo "$PROGNAME: $CONFIG_PREFIX already exists, aborted" >&2
+	echo >&2 $PROGNAME: "$CONFIG_PREFIX already exists, aborted"
+	exit 1
+fi
+if [ -z "$BACKEND_TYPE" ]; then
+	echo >&2 $PROGNAME: "-b <backend_type> option is necessary, aborted"
 	exit 1
 fi
 
-. $prefix/share/gfarm/config/config-gfarm.$BACKEND_TYPE
+PATH="$prefix/bin:/bin:/sbin:/usr/bin:/usr/sbin:/usr/ucb:/usr/pkg/bin:/usr/pkg/sbin:/usr/local/bin:/usr/local/sbin:$PATH"
+export PATH
+awk=awk
+if [ -f /usr/bin/nawk ]; then awk=/usr/bin/nawk; fi
+
+config_dir="${prefix}/share/gfarm/config"
+. $config_dir/config-gfarm.$BACKEND_TYPE
+. $config_dir/config-gfarm.sysdep
 set_first_defaults_$BACKEND_TYPE
 
 # sysdep_defaults must set: $RC_DIR
 sysdep_defaults
 
-PATH=$prefix/bin:$PATH 
-export PATH
-
-BACKEND_PORT=50602
-GFMD_PORT=50601
-GFSD_PORT=50600
-
 set_last_defaults_$BACKEND_TYPE
 
-config-gfarm --prefix $CONFIG_PREFIX -b $BACKEND_TYPE \
+config-gfarm --prefix $CONFIG_PREFIX -b $BACKEND_TYPE $hostname \
 	-p $BACKEND_PORT -m $GFMD_PORT -s $GFSD_PORT 
-config-gfsd --prefix $CONFIG_PREFIX -l `hostname`
+config-gfsd --prefix $CONFIG_PREFIX -l `hostname` $hostname
 config-gfsd --prefix $CONFIG_PREFIX -l localhost -h localhost
 
 GFARM_CONFIG_FILE=$CONFIG_PREFIX/etc/gfarm.conf
@@ -54,15 +103,15 @@ export GFARM_CONFIG_FILE
 gfmkdir '~'
 
 case `uname` in
-Linux)	env LD_PRELOAD="$prefix/lib/libgfs_hook.so.0 /usr/lib/gfarm/librt-not-hidden.so /usr/lib/gfarm/libpthread-not-hidden.so /usr/lib/gfarm/libc-not-hidden.so" $regress;;
-SunOS)	env LD_PRELOAD_32=$prefix/lib/libgfs_hook.so.0 $regress;;
-HP-UX)	env LD_PRELOAD=$prefix/lib/libgfs_hook.sl $regress;;
-OSF1)	env _RLD_LIST="$prefix/lib/libgfs_hook.so.0:DEFAULT" $regress;;
-Darwin)	env DYLD_INSERT_LIBRARIES=$prefix/lib/libgfs_hook.dylib DYLD_FORCE_FLAT_NAMESPACE= $regress;;
+Linux)	env LD_PRELOAD="$prefix/lib/libgfs_hook.so.0 /usr/lib/gfarm/librt-not-hidden.so /usr/lib/gfarm/libpthread-not-hidden.so /usr/lib/gfarm/libc-not-hidden.so" "$@";;
+SunOS)	env LD_PRELOAD_32=$prefix/lib/libgfs_hook.so.0 "$@";;
+HP-UX)	env LD_PRELOAD=$prefix/lib/libgfs_hook.sl "$@";;
+OSF1)	env _RLD_LIST="$prefix/lib/libgfs_hook.so.0:DEFAULT" "$@";;
+Darwin)	env DYLD_INSERT_LIBRARIES=$prefix/lib/libgfs_hook.dylib DYLD_FORCE_FLAT_NAMESPACE= "$@";;
 NetBSD|FreeBSD)
-	env LD_PRELOAD=$prefix/lib/libgfs_hook.so.0 $regress;;
-*)	echo "WARNING: unknown OS type `uname`. gfs_hook test will fail." >&2
-	env LD_PRELOAD=$prefix/lib/libgfs_hook.so.0 $regress;;
+	env LD_PRELOAD=$prefix/lib/libgfs_hook.so.0 "$@";;
+*)	echo >&2 "WARNING: unknown OS type `uname`. gfs_hook test may fail."
+	env LD_PRELOAD=$prefix/lib/libgfs_hook.so.0 "$@";;
 esac
 
 
@@ -70,4 +119,7 @@ service_ctl gfsd-localhost stop
 service_ctl gfsd-`hostname` stop
 service_ctl gfmd stop
 service_stop_$BACKEND_TYPE
-rm -rf $CONFIG_PREFIX
+if $remove_environment; then
+	rm -rf $CONFIG_PREFIX
+fi
+
